refactor: divide client package
