+++ batch/src/main/java/nexters/payout/batch/application/client/StockLogo.java
@@ -1,4 +1,4 @@
-package nexters.payout.batch.application;
+package nexters.payout.batch.application.client;
 
 public interface StockLogo {
     String getLogoUrl(String ticker);
