+++ batch/src/test/java/nexters/payout/batch/common/AbstractBatchServiceTest.java
@@ -1,9 +1,11 @@
 package nexters.payout.batch.common;
 
 import nexters.payout.batch.application.DividendBatchService;
-import nexters.payout.batch.application.FinancialClient;
+import nexters.payout.batch.application.PortfolioBatchService;
+import nexters.payout.batch.application.client.FinancialClient;
 import nexters.payout.batch.application.StockBatchService;
 import nexters.payout.domain.dividend.domain.repository.DividendRepository;
+import nexters.payout.domain.portfolio.domain.repository.PortfolioRepository;
 import nexters.payout.domain.stock.domain.repository.StockRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,13 +26,20 @@ public abstract class AbstractBatchServiceTest {
     public DividendRepository dividendRepository;
 
     @Autowired
+    public PortfolioRepository portfolioRepository;
+
+    @Autowired
     public StockBatchService stockBatchService;
 
     @Autowired
     public DividendBatchService dividendBatchService;
 
+    @Autowired
+    public PortfolioBatchService portfolioBatchService;
+
     @AfterEach
     void afterEach() {
+        portfolioRepository.deleteAll();
         dividendRepository.deleteAll();
         stockRepository.deleteAll();
     }
